[flang] Removed obsolete comments and conversions
The addition of a more robust Integer constructor for POD types has
made some comments and static_cast expressions obsolete.

Original-commit: flang-compiler/f18@847b8a768417bb3ccfa5dd919364ac39880bbb5e
Reviewed-on: https://github.com/flang-compiler/f18/pull/208

+++ flang/lib/evaluate/expression.cc
@@ -476,10 +476,8 @@ template<int KIND>
 Expr<SubscriptInteger> Expr<Type<TypeCategory::Character, KIND>>::LEN() const {
   return std::visit(
       common::visitors{[](const Constant<Result> &c) {
-                         // std::string::size_type isn't convertible to uint64_t
-                         // on Darwin
-                         return AsExpr(Constant<SubscriptInteger>{
-                             static_cast<std::uint64_t>(c.value.size())});
+                         return AsExpr(
+                             Constant<SubscriptInteger>{c.value.size()});
                        },
           [](const Parentheses<Result> &x) { return x.left().LEN(); },
           [](const Concat<KIND> &c) {
